# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Deep Control Loads Backend Rest API Documentation",
        "description": "Rest API Documentation",
        "contact": {
            "email": "remstroyod@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register User Rest API",
                "description": "Register User",
                "operationId": "auth.register",
                "requestBody": {
                    "description": "Create user object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "User Name",
                                        "type": "string",
                                        "example": "Name"
                                    },
                                    "surname": {
                                        "description": "User Surname",
                                        "type": "string",
                                        "example": "Surname"
                                    },
                                    "email": {
                                        "description": "User Email",
                                        "type": "string",
                                        "example": "admin@example.com"
                                    },
                                    "password": {
                                        "description": "User Password",
                                        "type": "string",
                                        "example": 123456
                                    },
                                    "password_confirmation": {
                                        "description": "User Password Confirmation",
                                        "type": "string",
                                        "example": 123456
                                    },
                                    "position_id": {
                                        "description": "User Position",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "gender": {
                                        "description": "User Gender",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "company_name": {
                                        "description": "User Company",
                                        "type": "string",
                                        "example": "My Company"
                                    },
                                    "street": {
                                        "description": "User Street",
                                        "type": "string",
                                        "example": "Street Name"
                                    },
                                    "post": {
                                        "description": "User Post",
                                        "type": "string",
                                        "example": "65000"
                                    },
                                    "city": {
                                        "description": "User City",
                                        "type": "string",
                                        "example": "Odessa"
                                    },
                                    "country_id": {
                                        "description": "User Country",
                                        "type": "integer",
                                        "example": "Odessa"
                                    },
                                    "salutation": {
                                        "description": "User Salutation",
                                        "type": "string",
                                        "example": "Salutation"
                                    },
                                    "phone": {
                                        "description": "User Phone",
                                        "type": "string",
                                        "example": "+380960000000"
                                    },
                                    "confirm_docs": {
                                        "description": "User Confirm Docs",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "subcontractors": {
                                        "description": "User Subcontractors",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "trailers": {
                                        "description": "User Trailers",
                                        "type": "object",
                                        "example": {
                                            "1": 2,
                                            "2": 4
                                        },
                                        "additionalProperties": {
                                            "description": "The Trailer ID as key and Count as value",
                                            "type": "integer"
                                        }
                                    },
                                    "tractors": {
                                        "description": "User Tractors",
                                        "type": "object",
                                        "example": {
                                            "1": 1,
                                            "2": 3,
                                            "3": 6
                                        },
                                        "additionalProperties": {
                                            "description": "The Tractors ID as key and Count as value",
                                            "type": "integer"
                                        }
                                    },
                                    "miscellaneous": {
                                        "description": "User Miscellaneous",
                                        "type": "object",
                                        "example": {
                                            "1": 7
                                        },
                                        "additionalProperties": {
                                            "description": "The Miscellaneous ID as key and Count as value",
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/register/upload-file": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register File Upload User Rest API",
                "description": "Register File Upload User",
                "operationId": "auth.register.upload",
                "requestBody": {
                    "description": "Create user object",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "description": "File Passport",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "filename": {
                                        "description": "File Type",
                                        "type": "string",
                                        "example": "passport or cmr or license"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login User Rest API",
                "description": "Login User",
                "operationId": "auth.login",
                "requestBody": {
                    "description": "Create user object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "User Email",
                                        "type": "string",
                                        "example": "admin@example.com"
                                    },
                                    "password": {
                                        "description": "User Password",
                                        "type": "string",
                                        "example": 123456
                                    },
                                    "remember": {
                                        "description": "User Remember Me",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout User Rest API",
                "description": "Logout User",
                "operationId": "auth.logout",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/register/check-user": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Check User Rest API",
                "description": "Check User",
                "operationId": "auth.register.checkuser",
                "requestBody": {
                    "description": "Create user object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "User Email",
                                        "type": "string",
                                        "example": "admin@example.com"
                                    },
                                    "password": {
                                        "description": "User Password",
                                        "type": "string",
                                        "example": 123456
                                    },
                                    "password_confirmation": {
                                        "description": "User Password Confirmation",
                                        "type": "string",
                                        "example": 123456
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "A short description for security scheme",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    }
}